removed all QT3 code. some crashing in system evaluation dialog now. Rig settings need more work.

+++ common/datadecoding/MOTSlideShow.h
@@ -65,7 +65,7 @@ class CMOTSlideShowEncoder
     CMOTDABEnc MOTDAB;
 
     vector <SPicDescr> vecPicFileNames;
-    int iPictureCnt;
+    size_t iPictureCnt;
 
     string strCurObjName;
 };
